requirements.txt
# Discord Bot
discord.py>=2.3.0

# Environment Management
python-dotenv>=1.0.0

# AI/ML
google-generativeai>=0.3.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utilities
aiohttp>=3.9.0

